Fixing crashes for ScanlineFill. Still not working as intended

+++ ExList2/ScanlineFill/drawingarea.h
@@ -12,8 +12,8 @@
 
 struct edge {
              int maxY, minY;
-             int currentX;
-             int xIncr;
+             double currentX;
+             double xIncr;
              };
 
 struct compare
